+++ src/main/java/phankhanh/book_store/DTO/request/ReqBookUpdate.java
@@ -1,6 +1,8 @@
 package phankhanh.book_store.DTO.request;
 
 import jakarta.validation.constraints.*;
+
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Set;
@@ -28,8 +30,8 @@ public record ReqBookUpdate(
         @NotBlank String ageRating,
 
         @NotBlank String status,
-        @NotNull @Positive Long price,
-        @Positive Long salePrice,
+        @NotNull @Positive BigDecimal price,
+        @Positive BigDecimal salePrice,
         Instant saleStartAt,
         Instant saleEndAt,
